Move to Subcomponents No ACTION_READLN
Last typo fix on timers, and a full reset for modernization - all ACTION_READLN moved to optional subcomponents. Need to keep DESIGNATED static now, but plentry of space. Note: need to set up config.ini for advanced users and map it as full functionality is still in place, just using a tiny fraction of the combinations.